move and rename the factory method
