+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/routing_policy/communities/CommunitySetExprs.java
@@ -1,6 +1,10 @@
 package org.batfish.datamodel.routing_policy.communities;
 
+import dk.brics.automaton.Automaton;
+import dk.brics.automaton.RegExp;
 import javax.annotation.Nonnull;
+import org.apache.commons.text.StringEscapeUtils;
+import org.batfish.datamodel.routing_policy.Common;
 
 /** Utility class for constructing {@link CommunitySetExpr}s. */
 public final class CommunitySetExprs {
@@ -12,4 +16,34 @@ public final class CommunitySetExprs {
   private static final CommunitySetExpr EMPTY = new LiteralCommunitySet(CommunitySet.empty());
 
   private CommunitySetExprs() {}
+
+  /**
+   * Creates an expression that matches on the given community-set regex. If the regex can be
+   * identified as one that only requires the existence of a single community, then the result makes
+   * that explicit by using {@link CommunityMatchRegex} instead of {@link CommunitySetMatchRegex}.
+   *
+   * @param regex the regex
+   * @return an expression that represents a community-set match on the regex
+   */
+  public static @Nonnull CommunitySetMatchExpr fromRegex(String regex) {
+    String trimmedRegex = regex;
+    // a conservative check to determine if the regex only matches on the existence of a single
+    // community in the set: the regex optionally starts with _, optionally ends with _, and in
+    // between only accepts strings containing digits and colons
+    String underscore = StringEscapeUtils.unescapeJava(Common.DEFAULT_UNDERSCORE_REPLACEMENT);
+    if (trimmedRegex.startsWith(underscore)) {
+      trimmedRegex = trimmedRegex.substring(underscore.length());
+    }
+    if (trimmedRegex.endsWith(underscore)) {
+      trimmedRegex = trimmedRegex.substring(0, trimmedRegex.length() - underscore.length());
+    }
+    Automaton trimmedRegexAuto = new RegExp(trimmedRegex).toAutomaton();
+    Automaton digitsAndColons = new RegExp("[0-9:]+").toAutomaton();
+    if (trimmedRegexAuto.intersection(digitsAndColons).equals(trimmedRegexAuto)) {
+      return new HasCommunity(new CommunityMatchRegex(ColonSeparatedRendering.instance(), regex));
+    } else {
+      return new CommunitySetMatchRegex(
+          new TypesFirstAscendingSpaceSeparated(ColonSeparatedRendering.instance()), regex);
+    }
+  }
 }
