+++ projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/routing_policy/communities/CommunitySetExprsTest.java
@@ -1,27 +1,23 @@
-package org.batfish.datamodel.routing_policy;
+package org.batfish.datamodel.routing_policy.communities;
 
-import static org.batfish.datamodel.routing_policy.Common.communitySetMatchRegex;
+import static org.batfish.datamodel.routing_policy.communities.CommunitySetExprs.fromRegex;
 import static org.junit.Assert.assertEquals;
 
 import org.apache.commons.text.StringEscapeUtils;
-import org.batfish.datamodel.routing_policy.communities.ColonSeparatedRendering;
-import org.batfish.datamodel.routing_policy.communities.CommunityMatchRegex;
-import org.batfish.datamodel.routing_policy.communities.CommunitySetMatchRegex;
-import org.batfish.datamodel.routing_policy.communities.HasCommunity;
-import org.batfish.datamodel.routing_policy.communities.TypesFirstAscendingSpaceSeparated;
+import org.batfish.datamodel.routing_policy.Common;
 import org.junit.Test;
 
-public final class CommonTest {
+public final class CommunitySetExprsTest {
 
   @Test
   public void testCommunitySetMatchRegexUnoptimized() {
     assertEquals(
-        communitySetMatchRegex("^65000:123 65011:12[3]$"),
+        fromRegex("^65000:123 65011:12[3]$"),
         new CommunitySetMatchRegex(
             new TypesFirstAscendingSpaceSeparated(ColonSeparatedRendering.instance()),
             "^65000:123 65011:12[3]$"));
     assertEquals(
-        communitySetMatchRegex("^$"),
+        fromRegex("^$"),
         new CommunitySetMatchRegex(
             new TypesFirstAscendingSpaceSeparated(ColonSeparatedRendering.instance()), "^$"));
   }
@@ -32,27 +28,27 @@ public final class CommonTest {
     String underscore = StringEscapeUtils.unescapeJava(Common.DEFAULT_UNDERSCORE_REPLACEMENT);
 
     assertEquals(
-        communitySetMatchRegex("53"),
+        fromRegex("53"),
         new HasCommunity(new CommunityMatchRegex(ColonSeparatedRendering.instance(), "53")));
     assertEquals(
-        communitySetMatchRegex("53:"),
+        fromRegex("53:"),
         new HasCommunity(new CommunityMatchRegex(ColonSeparatedRendering.instance(), "53:")));
     assertEquals(
-        communitySetMatchRegex(underscore + "53:"),
+        fromRegex(underscore + "53:"),
         new HasCommunity(
             new CommunityMatchRegex(ColonSeparatedRendering.instance(), underscore + "53:")));
     assertEquals(
-        communitySetMatchRegex(":53"),
+        fromRegex(":53"),
         new HasCommunity(new CommunityMatchRegex(ColonSeparatedRendering.instance(), ":53")));
     assertEquals(
-        communitySetMatchRegex(":53" + underscore),
+        fromRegex(":53" + underscore),
         new HasCommunity(
             new CommunityMatchRegex(ColonSeparatedRendering.instance(), ":53" + underscore)));
     assertEquals(
-        communitySetMatchRegex("[0-9]+:"),
+        fromRegex("[0-9]+:"),
         new HasCommunity(new CommunityMatchRegex(ColonSeparatedRendering.instance(), "[0-9]+:")));
     assertEquals(
-        communitySetMatchRegex("[0-9]+:[123]*"),
+        fromRegex("[0-9]+:[123]*"),
         new HasCommunity(
             new CommunityMatchRegex(ColonSeparatedRendering.instance(), "[0-9]+:[123]*")));
   }
